Fixed story continuity
FIRST OF ALL i was wrong, it wasn't the scrap gun. it was the music box. but they're in the same room so whatever

basically mutant dinky appears in the music box pickup dialogue but you can see this before getting to mutant dinky in the story. i've taken out the one line in the dialogue to fix the spoiler. scrap gun dialogue is fine afaik